requirements.txt
fastapi
uvicorn
python-dotenv
itsdangerous
jinja2
python-multipart
qrcode[pil]
